feat: implement notification and conversation features with comprehensive tests
- Added `useNotificationCount` hook to manage notification counts with optional initial values.
- Created `getConversationList`, `getConversationMeta`, and `getInitialMessages` functions for conversation management.
- Developed `getFriendRequestsData` and `getNotificationCounts` functions to handle friend requests and notification counts.
- Introduced new data models and types for conversations, notifications, and user profiles.
- Implemented tests for conversations, friends, notifications, profiles, sessions, and users to ensure functionality and reliability.
- Added `server-only` package to manage server-side only imports.
- Updated package dependencies and lock files accordingly.
