+++ app/api/friend-requests/__tests__/route.test.ts
@@ -80,40 +80,3 @@ describe("POST /api/friend-requests", () => {
 		expect(res.status).toBe(201);
 	});
 });
-
-describe("GET /api/friend-requests", () => {
-	it("returns 401 when not authenticated", async () => {
-		setupApiMocks(testDb, null);
-		const { GET } = await import("@/app/api/friend-requests/route");
-		const res = await GET();
-		expect(res.status).toBe(401);
-	});
-
-	it("returns empty arrays when no requests", async () => {
-		setupApiMocks(testDb, "user-1");
-		const { GET } = await import("@/app/api/friend-requests/route");
-		const res = await GET();
-		const data = await res.json();
-		expect(data.received).toEqual([]);
-		expect(data.sent).toEqual([]);
-		expect(data.accepted).toEqual([]);
-	});
-
-	it("categorizes requests correctly", async () => {
-		seedFriendRequest(testDb, "fr-1", "user-2", "user-1", "pending");
-		seedFriendRequest(testDb, "fr-2", "user-1", "user-3", "pending");
-		seedFriendRequest(testDb, "fr-3", "user-1", "user-2", "accepted");
-
-		setupApiMocks(testDb, "user-1");
-		const { GET } = await import("@/app/api/friend-requests/route");
-		const res = await GET();
-		const data = await res.json();
-
-		expect(data.received).toHaveLength(1);
-		expect(data.received[0].id).toBe("fr-1");
-		expect(data.sent).toHaveLength(1);
-		expect(data.sent[0].id).toBe("fr-2");
-		expect(data.accepted).toHaveLength(1);
-		expect(data.accepted[0].id).toBe("fr-3");
-	});
-});
