+++ app/api/friend-requests/route.ts
@@ -99,73 +99,3 @@ export async function POST(request: Request) {
 		{ status: 201 },
 	);
 }
-
-export async function GET() {
-	const auth = await requireAuth();
-	if (!auth) return unauthorized();
-	const { userId } = auth;
-
-	const requests = db
-		.select()
-		.from(friendRequest)
-		.where(
-			or(
-				eq(friendRequest.senderId, userId),
-				eq(friendRequest.receiverId, userId),
-			),
-		)
-		.all();
-
-	if (requests.length === 0) {
-		return Response.json({ received: [], sent: [], accepted: [] });
-	}
-
-	const userIds = [
-		...new Set(requests.flatMap((r) => [r.senderId, r.receiverId])),
-	];
-
-	const users = db
-		.select({
-			id: user.id,
-			name: user.name,
-			username: user.username,
-			image: user.image,
-		})
-		.from(user)
-		.where(or(...userIds.map((uid) => eq(user.id, uid))))
-		.all();
-
-	const userMap = new Map(users.map((u) => [u.id, u]));
-
-	const received: Array<Record<string, unknown>> = [];
-	const sent: Array<Record<string, unknown>> = [];
-	const accepted: Array<Record<string, unknown>> = [];
-
-	for (const r of requests) {
-		const sender = userMap.get(r.senderId) ?? null;
-		const receiverUser = userMap.get(r.receiverId) ?? null;
-		const friendUser = r.senderId === userId ? receiverUser : sender;
-
-		const entry = {
-			id: r.id,
-			senderId: r.senderId,
-			receiverId: r.receiverId,
-			status: r.status,
-			createdAt: r.createdAt?.getTime() ?? null,
-			updatedAt: r.updatedAt?.getTime() ?? null,
-			sender,
-			receiver: receiverUser,
-			friend: friendUser,
-		};
-
-		if (r.status === "accepted") {
-			accepted.push(entry);
-		} else if (r.status === "pending" && r.receiverId === userId) {
-			received.push(entry);
-		} else if (r.status === "pending" && r.senderId === userId) {
-			sent.push(entry);
-		}
-	}
-
-	return Response.json({ received, sent, accepted });
-}
