+++ components/feature/profile-actions.tsx
@@ -1,0 +1,160 @@
+"use client";
+
+import { Pencil } from "lucide-react";
+import { useRouter } from "next/navigation";
+import { useCallback, useState } from "react";
+import { FriendRequestButton } from "@/components/feature/friend-request-button";
+import { Button } from "@/components/ui/button";
+import {
+	Dialog,
+	DialogContent,
+	DialogDescription,
+	DialogFooter,
+	DialogHeader,
+	DialogTitle,
+	DialogTrigger,
+} from "@/components/ui/dialog";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Textarea } from "@/components/ui/textarea";
+import type { Profile } from "@/lib/types";
+
+interface ProfileActionsProps {
+	profile: Profile;
+}
+
+export function ProfileActions({ profile }: ProfileActionsProps) {
+	const router = useRouter();
+	const [dialogOpen, setDialogOpen] = useState(false);
+	const [saving, setSaving] = useState(false);
+
+	const [editName, setEditName] = useState("");
+	const [editLastName, setEditLastName] = useState("");
+	const [editBio, setEditBio] = useState("");
+	const [editImage, setEditImage] = useState("");
+
+	const displayName = profile.username || profile.name;
+
+	const openEdit = useCallback(() => {
+		setEditName(profile.name);
+		setEditLastName(profile.lastName ?? "");
+		setEditBio(profile.bio ?? "");
+		setEditImage(profile.image ?? "");
+		setDialogOpen(true);
+	}, [profile]);
+
+	const handleSave = useCallback(async () => {
+		setSaving(true);
+		try {
+			const res = await fetch("/api/profile", {
+				method: "PATCH",
+				headers: { "Content-Type": "application/json" },
+				body: JSON.stringify({
+					name: editName,
+					lastName: editLastName || null,
+					bio: editBio || null,
+					image: editImage || null,
+				}),
+			});
+			if (res.ok) {
+				setDialogOpen(false);
+				// Re-render the server component so the page shows fresh data
+				router.refresh();
+			}
+		} catch {
+			// Silently fail
+		} finally {
+			setSaving(false);
+		}
+	}, [editName, editLastName, editBio, editImage, router]);
+
+	if (!profile.isOwnProfile) {
+		return (
+			<FriendRequestButton
+				userId={displayName}
+				initialStatus={profile.friendStatus}
+				requestId={profile.friendRequestId ?? undefined}
+			/>
+		);
+	}
+
+	return (
+		<Dialog open={dialogOpen} onOpenChange={setDialogOpen}>
+			<DialogTrigger asChild>
+				<Button variant="outline" onClick={openEdit}>
+					<Pencil className="size-4" />
+					Edit profile
+				</Button>
+			</DialogTrigger>
+			<DialogContent>
+				<DialogHeader>
+					<DialogTitle>Edit profile</DialogTitle>
+					<DialogDescription>
+						Update your profile information. Username and email cannot be
+						changed.
+					</DialogDescription>
+				</DialogHeader>
+				<div className="space-y-4">
+					<div>
+						<Label>Username</Label>
+						<Input value={profile.username} disabled className="mt-1" />
+					</div>
+					<div className="grid grid-cols-2 gap-3">
+						<div>
+							<Label htmlFor="edit-name">First name</Label>
+							<Input
+								id="edit-name"
+								value={editName}
+								onChange={(e) => setEditName(e.target.value)}
+								placeholder="Your first name"
+								className="mt-1"
+								maxLength={50}
+							/>
+						</div>
+						<div>
+							<Label htmlFor="edit-lastname">Last name</Label>
+							<Input
+								id="edit-lastname"
+								value={editLastName}
+								onChange={(e) => setEditLastName(e.target.value)}
+								placeholder="Your last name"
+								className="mt-1"
+								maxLength={50}
+							/>
+						</div>
+					</div>
+					<div>
+						<Label htmlFor="edit-bio">Bio</Label>
+						<Textarea
+							id="edit-bio"
+							value={editBio}
+							onChange={(e) => setEditBio(e.target.value)}
+							placeholder="Tell us about yourself..."
+							className="mt-1"
+							maxLength={200}
+							rows={3}
+						/>
+						<span className="mt-1 block text-right text-xs text-muted-foreground">
+							{editBio.length}/200
+						</span>
+					</div>
+					<div>
+						<Label htmlFor="edit-image">Profile picture URL</Label>
+						<Input
+							id="edit-image"
+							value={editImage}
+							onChange={(e) => setEditImage(e.target.value)}
+							placeholder="https://example.com/avatar.jpg"
+							className="mt-1"
+						/>
+					</div>
+				</div>
+				<DialogFooter>
+					<Button onClick={handleSave} disabled={saving || !editName.trim()}>
+						{saving ? "Saving..." : "Save"}
+					</Button>
+				</DialogFooter>
+			</DialogContent>
+		</Dialog>
+	);
+}
