+++ hooks/use-chat-list.ts
@@ -1,54 +1,39 @@
 "use client";
 
+import { useRouter } from "next/navigation";
 import { useCallback, useEffect, useState } from "react";
 import { SocketEvent } from "@/lib/constants";
 import type { Conversation } from "@/lib/types";
 import { useSocket } from "./use-socket";
 
-interface GroupedConversations {
+export interface GroupedConversations {
 	friends: Conversation[];
 	others: Conversation[];
 }
 
-export function useChatList() {
-	const [conversations, setConversations] = useState<GroupedConversations>({
-		friends: [],
-		others: [],
-	});
-	const [isLoading, setIsLoading] = useState(true);
+export function useChatList(initialData: GroupedConversations) {
+	const router = useRouter();
+	const [conversations, setConversations] =
+		useState<GroupedConversations>(initialData);
 	const { socket } = useSocket();
 
-	const fetchConversations = useCallback(async () => {
-		try {
-			const res = await fetch("/api/conversations");
-			if (res.ok) {
-				const data = await res.json();
-				if (Array.isArray(data)) {
-					const friends = data.filter((c: Conversation) => c.isFriend);
-					const others = data.filter((c: Conversation) => !c.isFriend);
-					setConversations({ friends, others });
-				} else {
-					setConversations({
-						friends: data.friends ?? [],
-						others: data.others ?? [],
-					});
-				}
-			}
-		} catch {
-			// Silently fail; user can reload
-		} finally {
-			setIsLoading(false);
-		}
-	}, []);
+	// Server data is authoritative: re-sync whenever the server
+	// component re-renders (router.refresh() after a socket event).
+	useEffect(() => {
+		setConversations(initialData);
+	}, [initialData]);
+
+	const refetch = useCallback(() => {
+		router.refresh();
+	}, [router]);
 
 	useEffect(() => {
-		fetchConversations();
 		function onFocus() {
-			fetchConversations();
+			refetch();
 		}
 		window.addEventListener("focus", onFocus);
 		return () => window.removeEventListener("focus", onFocus);
-	}, [fetchConversations]);
+	}, [refetch]);
 
 	useEffect(() => {
 		if (!socket) return;
@@ -63,13 +48,13 @@ export function useChatList() {
 				const allConvos = [...prev.friends, ...prev.others];
 				const idx = allConvos.findIndex((c) => c.id === data.conversationId);
 				if (idx === -1) {
-					fetchConversations();
+					// Unknown conversation: let the server rebuild the list
+					refetch();
 					return prev;
 				}
 
 				const conv = allConvos[idx];
-				const prevUnread =
-					(conv as Conversation & { unreadCount?: number }).unreadCount ?? 0;
+				const prevUnread = conv.unreadCount ?? 0;
 
 				const updated = {
 					...conv,
@@ -91,7 +76,7 @@ export function useChatList() {
 		}
 
 		function handleFriendUpdate() {
-			fetchConversations();
+			refetch();
 		}
 
 		socket.on(SocketEvent.MESSAGE_NEW, handleNewMessage);
@@ -100,7 +85,7 @@ export function useChatList() {
 			socket.off(SocketEvent.MESSAGE_NEW, handleNewMessage);
 			socket.off(SocketEvent.FRIEND_REQUEST_UPDATE, handleFriendUpdate);
 		};
-	}, [socket, fetchConversations]);
+	}, [socket, refetch]);
 
-	return { conversations, isLoading, refetch: fetchConversations };
+	return { conversations, refetch };
 }
