+++ package.json
@@ -33,6 +33,7 @@
 		"react": "19.2.4",
 		"react-dom": "19.2.4",
 		"react-hook-form": "^7.74.0",
+		"server-only": "^0.0.1",
 		"shadcn": "^4.6.0",
 		"socket.io": "^4.8.3",
 		"socket.io-client": "^4.8.3",
